requirements.txt
Django==2.2
djangorestframework==3.9.2
Pygments==2.3.1
pytz==2018.9
sqlparse==0.3.0
